requirements.txt
asgiref==3.7.2
Django==4.1.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
Pygments==2.15.1
PyJWT==2.7.0
pymongo==4.3.3
python-dotenv==1.0.0
python-snappy==0.6.1
pytz==2023.3
sqlparse==0.2.4
typing_extensions==4.6.3
tzdata==2023.3